add hibernate

+++ src/main/java/com/example/foxstudent105614/service/ReportSchoolService.java
@@ -3,23 +3,30 @@ package com.example.foxstudent105614.service;
 import com.example.foxstudent105614.model.Course;
 import com.example.foxstudent105614.model.Group;
 import com.example.foxstudent105614.model.Student;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class ReportSchoolService {
+    private final Logger logger = LoggerFactory.getLogger(ReportSchoolService.class);
+
     private final SchoolService schoolService;
 
     public ReportSchoolService(SchoolService schoolService) {
         this.schoolService = schoolService;
+
     }
 
     public void printGroupsByStudentCount(int maxStudentCount) {
         List<Group> groups = schoolService.findGroupsByStudentCount(maxStudentCount);
         if (groups.isEmpty()) {
-            System.out.println("No groups found with " + maxStudentCount + " or fewer students.");
+            logger.info("No groups found with {} or fewer students.", maxStudentCount);
         } else {
-            System.out.println("Groups of " + maxStudentCount + " or fewer students:");
+            logger.info("Groups of {} or fewer students:", maxStudentCount);
             for (Group group : groups) {
                 System.out.println(group.toString());
             }
@@ -29,23 +36,19 @@ public class ReportSchoolService {
     public void printStudentsByCourseName(String courseName) {
         List<Student> students = schoolService.findStudentsByCourseName(courseName);
         if (students.isEmpty()) {
-            System.out.println("No students found for the course with name: " + courseName);
+            logger.info("No students found for the course with name: {}", courseName);
         } else {
-            System.out.println("Students related to the course '" + courseName + "':");
+            logger.info("Students related to the course '{}':", courseName);
             for (Student student : students) {
-                System.out.println(student.firstName() + " " + student.lastName() + " (ID: " + student.studentId() + ")");
+                logger.info("{} {} (ID: {})", student.getFirstName(), student.getLastName(), student.getStudentId());
             }
         }
     }
 
-    public void printAddStudent(int groupId, String firstName, String lastName) {
-        Optional<Group> groupOptional = schoolService.findGroupById(groupId);
-
-        if (groupOptional.isEmpty()) {
-            System.out.println("Error: Group not found for id: " + groupId);
-        } else {
-            schoolService.addStudent(groupId, firstName, lastName);
-        }
+    public void printAddStudent(String firstName, String lastName) {
+        Student student = new Student(firstName, lastName);
+        schoolService.addStudent(student);
+        logger.info("Added new student. First name: {}, Last name: {}", firstName, lastName);
     }
 
     public void printDeleteStudentById(int studentId) {
@@ -53,37 +56,41 @@ public class ReportSchoolService {
 
         studentOptional.ifPresent(student -> {
             schoolService.deleteStudentById(studentId);
-            System.out.println("Deleted student with ID: " + studentId);
+            logger.info("Deleted student: {}", student);
         });
 
         if (studentOptional.isEmpty()) {
-            System.out.println("Student not found with ID: " + studentId);
+            logger.error("Student not found with ID: {}", studentId);
         }
     }
 
-    public void printCoursesForStudent(int studentId, List<Integer> courseListId) {
+    public void printSaveStudentInCourse(int studentId, int courseId) {
         Optional<Student> studentOptional = schoolService.findStudentById(studentId);
+        Optional<Course> courseOptional = schoolService.findCourseById(courseId);
 
         if (studentOptional.isPresent()) {
             Student student = studentOptional.get();
-            courseListId.forEach(courseId -> {
-                Optional<Course> courseOptional = schoolService.findCourseById(courseId);
 
-                if (courseOptional.isPresent()) {
-                    schoolService.saveStudentInCourse(studentId, courseId);
-                    System.out.println("Added student with ID " + student.studentId() + " to course with ID: " + courseId);
-                } else {
-                    System.out.println("Error: Course not found for ID - CourseID: " + courseId);
-                }
-            });
+            if (courseOptional.isPresent()) {
+                schoolService.saveStudentInCourse(studentOptional.get().getStudentId(), courseOptional.get().getCourseId());
+                logger.info("Added student {} {} to course with ID: {}", student.getFirstName(), student.getLastName(), courseId);
+            } else {
+                logger.error("Error: Course not found for ID - CourseID: {}", courseId);
+            }
         } else {
-            System.out.println("Error: Student not found for ID - StudentID: " + studentId);
+            logger.error("Error: Student not found for ID - StudentID: {}", studentId);
         }
     }
 
-    public void printRemoveStudentFromCourse(int studentId, int courseId) {
-        schoolService.deleteStudentFromCourse(studentId, courseId);
-        System.out.println("Removed student with ID " + studentId + " from course with ID " + courseId);
-    }
+    public void printDeleteStudentFromCourse(int studentId, int courseId) {
+        Optional<Student> studentOptional = schoolService.findStudentById(studentId);
+        Optional<Course> courseOptional = schoolService.findCourseById(courseId);
 
+        if (courseOptional.isPresent() && studentOptional.isPresent()) {
+            schoolService.deleteStudentFromCourse(studentId, courseId);
+            logger.info("Removed student with ID {} from course with ID {}", studentId, courseId);
+        } else {
+            logger.info("Course with ID {} or Student with ID {} not found", courseId, studentId);
+        }
+    }
 }
